+++ oremda/pipeline/engine/__init__.py
@@ -6,9 +6,9 @@ import signal
 
 from oremda.typing import ContainerType
 from oremda.clients.singularity import SingularityClient
-from oremda.pipeline.runner.rpc.client import RpcClient
-from oremda.pipeline.runner.context import pipeline_context
-from oremda.pipeline.runner.config import settings
+from oremda.pipeline.engine.rpc.client import RpcClient
+from oremda.pipeline.engine.context import pipeline_context
+from oremda.pipeline.engine.config import settings
 
 # Setup logger
 logger = logging.getLogger("engine")
